Extracted generic copy functions from GTestHelper
