+++ plugins/org.yakindu.sct.doc.user/.classpath
@@ -1,6 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <classpath>
-	<classpathentry kind="src" path="src/main/java"/>
 	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER"/>
 	<classpathentry kind="lib" path="lib/bin/com.google.guava/guava-15.0.jar"/>
 	<classpathentry kind="lib" path="lib/bin/org.jsoup/jsoup-1.7.2.jar"/>
