+++ test-plugins/org.yakindu.sct.generator.c.test/META-INF/MANIFEST.MF
@@ -21,4 +21,5 @@ Require-Bundle: org.junit,
  org.eclipse.xtext.ui.shared,
  org.yakindu.sct.domain.generic.resource
 Export-Package: org.yakindu.sct.generator.c.gtest,
- org.yakindu.sct.generator.c.test
+ org.yakindu.sct.generator.c.test,
+ org.yakindu.sct.generator.c.test.utils
